+++ src/pages/SignIn/Index.tsx
@@ -2,11 +2,12 @@ import React, { useState } from "react";
 import { Button, Typography, Modal } from "antd";
 import { FieldValues, useForm } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
-import { useNavigate } from "react-router-dom";
+import { useLocation, useNavigate } from "react-router-dom";
 import Field from "components/DefaultField/Index";
 import { useTranslation } from "react-i18next";
 import { PRIMARY } from "utils/colorConsts";
 import { PROFILE_ROUTE, SIGN_UP_ROUTE } from "utils/routeConsts";
+import useAuth from "hooks/useAuth";
 import S from "./style";
 import signInSchema from "./schema";
 
@@ -15,9 +16,20 @@ interface SignInForm extends FieldValues {
   password: string;
 }
 
+type LocationProps = {
+  state: {
+    from: Location;
+  };
+};
+
 export default function SignIn() {
-  const navigate = useNavigate();
+  const { signIn } = useAuth();
   const { t } = useTranslation();
+
+  const location = useLocation() as LocationProps;
+  const navigate = useNavigate();
+  const from = location.state?.from?.pathname || "/";
+
   const { Title, Text } = Typography;
   const [isErrorModalOpen, setIsErrorModalOpen] = useState(false);
   const { control, handleSubmit } = useForm<SignInForm>({
@@ -34,6 +46,10 @@ export default function SignIn() {
       onOk: () => navigate(PROFILE_ROUTE),
       centered: true,
     });
+
+    signIn();
+    // Navigate to the page the user was before being disconnected
+    navigate(from, { replace: true });
   };
 
   const error = () => {
