+++ src/utils/routeConsts.ts
@@ -7,4 +7,4 @@ export const SETTINGS_ROUTE = "/settings";
 export const PROFILE_ROUTE = "/profile";
 export const SIGN_UP_ROUTE = "/sign-up";
 export const COMPLETE_REGISTRATION_ROUTE = "/complete-registration";
-export const SIGN_IN = "/login";
+export const SIGN_IN_ROUTE = "/login";
